// src/models/DedicatedAnthosBaremetal.ts
/* WARNING: This file is auto-generated . Do not edit manually. */

import { DedicatedAnthosBaremetalStatusEnum } from './DedicatedAnthosBaremetalStatusEnum';

export type DedicatedAnthosBaremetal = {
  additional: boolean;
  id: string;
  name: string;
  privateIp: string;
  profile: string;
  rack: string;
  serviceName: string;
  stateful: boolean;
  status: DedicatedAnthosBaremetalStatusEnum;
};